beta-feat-ep-2.0-as-serviceregistry
Added support for utilising the EP 2.0 envs/messaging services as service registry. This provides support for Solace Cloud Services and manually configured services in EP. Env variable SERVICE_REGISTRY is used to switch on this behaviour. Current limitation - protocols and ports are hard coded to standard ports on Solace Software broker, SEMPv2 password is not resolved via the variables mechanism